Move features query filtering tests to a separate file
Refs AHTI-223
